split utilities into header+source files, fixed segfault on error=NULL
I should have split this into two commits but in my earnestness fixed a bug that was all across GobboVector where functions would write to the error parameter even if it was NULL
